add a `RefineFlavor` component
This allows us to use it as:

```html
<refine-flavor as="ul" component="path.to.component" :flavorOptions="{ isSelected: true }">
  ...
</refine-flavor>
```

- The `as` prop is the default component to use
- The `component` prop allows us to define a path to the component in
  the `flavor` override.
- The `:flavorOptions` is passed to the `class` and `style` options in a
  flavor override. This also means that the `class` and `style` options
  can be functions and they will always receive an object (your
  `flavorOptions`) or `{}` if no `:flavorOptions` are present.

You can also put any prop on the `<refine-flavor>` which will be put on
the actual component (The `as` if no `component` is provided in the
falvor override)

API naming is not set in stone so this can be changed if there is a
better name for it.

Also the location of the files is not perfect yet, but want to refactor
that at a later point in time.