manually passthrough attributes
If we render the `refine-date-picker` via the `refine-flavor` component,
then Vue can't figure out that it should forward the `attrs` to the
first "DOM" component, not the first component itself.

In this case, we want the attributs to end up in the
`refine-date-picker`, but they are put on the `refine-flavor` component
directly.

Luckily, we can bypass this automagicness by ignoring the
`inheritAttrs`, and forwarding it directly to the resolved component
from the flavor system.